requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pandas>=2.0.0
jinja2>=3.1.2
python-multipart>=0.0.6
